requirements.txt
crewai
crewai-tools
langchain-groq
langchain-google-genai
pandas
scikit-learn
python-dotenv